fixed company balance, now represent real numbers

+++ backend/Database/model.js
@@ -27,7 +27,7 @@ const Transaction = sequelize.define("transaction", {
 
 
 const Company = sequelize.define("company", {
-  balance: DataTypes.INTEGER,  
+  balance: DataTypes.DOUBLE,  
 })
 
 const Installment = sequelize.define("installment", {
